# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Retrieve a list of all users (Admin access required).",
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved users list.",
            "content": {
              "text/html": {
                "schema": {
                  "title": "Users",
                  "users": [
                    {
                      "id": 1,
                      "firstName": "John",
                      "lastName": "Doe",
                      "email": "john@example.com",
                      "role": "User"
                    },
                    {
                      "id": 2,
                      "firstName": "Jane",
                      "lastName": "Smith",
                      "email": "jane@example.com",
                      "role": "Admin"
                    }
                  ]
                }
              }
            }
          },
          "302": {
            "description": "Found"
          },
          "500": {
            "description": "Internal server error - Unable to fetch users."
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Retrieve details of a specific user (Admin or the user themselves).",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user details.",
            "content": {
              "text/html": {
                "schema": {
                  "title": "User Details",
                  "user": {
                    "id": 1,
                    "firstName": "John",
                    "lastName": "Doe",
                    "email": "john@example.com",
                    "role": "User"
                  }
                }
              }
            }
          },
          "302": {
            "description": "Found"
          },
          "404": {
            "description": "User not found.",
            "content": {
              "text/html": {
                "schema": {
                  "title": "Error",
                  "message": "User Not Found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - Failed to retrieve user details.",
            "content": {
              "text/html": {
                "schema": {
                  "title": "Error",
                  "message": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "Delete a user (Admin access required).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "✅ User deleted successfully!"
                }
              }
            }
          },
          "400": {
            "description": "User ID is required.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "User ID is required."
                }
              }
            }
          },
          "403": {
            "description": "Admins cannot be deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "Admins cannot be deleted!"
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "User not found."
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - Failed to delete user.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "Failed to delete user."
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}